+++ root/include/GeigerSensor.h
@@ -1,25 +1,31 @@
-#ifndef GEIGERSENSOR_H 
-#define GEIGERSENSOR_H
+#ifndef GEIGERINTERRUPT_H
+#define GEIGERINTERRUPT_H
 
 #include "Sensor.h"
-#include <string.h>
-#include <Arduino.h>
-// need to changed to the correct ones
-#define RX 0
-#define TX 1
+#include "Arduino.h"
+
+#define GEIGER_PIN A2
+
+static volatile int count;
 
 class GeigerSensor : public Sensor {
+
     private:
-        String nameCompiled = "Geiger Counter";
-        String csvHeaderCompiled = "Geiger CPS, Geiger CPM, Geiger uSv/hr, Geiger Mode, ";
-    public:
+        String nameCompiled = "GeigerSensor";
+        String csvHeaderCompiled = "GeigerSensor CPS, ";
+        int measuringPeriodStart = 0; 
+        static void geigerCounter(){
+            count++;
+        }
+    public: 
         const String& getSensorName() const override;
-        const String& getSensorCSVHeader() const; 
+        const String& getSensorCSVHeader() const override;
         bool verifyPin() override;
         String readData() override;
-        String readEmpty() override {
-            return "-, -,  -, -,";
+        String readEmpty(){
+            return "-, ";
         }
+
 };
 
 #endif
