--- root/include/GeigerSlowSensor.h
@@ -1,28 +1,0 @@
-// Wind Sensor and temp 
-
-#ifndef GEIGERSLOWSENSOR_H 
-#define GEIGERSLOWSENSOR_H
-
-#include "Sensor.h"
-#include <string.h>
-#include <Arduino.h>
-// need to changed to the correct ones
-#define GEIGER_PIN A2
-#define MILLIS_BETWEEN_READINGS 1000 // 10 seconds
-
-class GeigerSlowSensor : public Sensor {
-    private:
-        int timer; 
-        String nameCompiled = "GeigerSlow";
-        String csvHeaderCompiled = "GeigerSlow CPS, ";
-    public:
-        const String& getSensorName() const override;
-        const String& getSensorCSVHeader() const; 
-        bool verifyPin() override;
-        String readData() override;
-        String readEmpty() override {
-            return "-, ";
-        }
-};
-
-#endif
